+++ packages/terra-avatar/src/index.js
@@ -1,0 +1,6 @@
+import Avatar from './variants/Avatar';
+import Facility from './variants/Facility';
+import SharedUser from './variants/SharedUser';
+
+export default Avatar;
+export { Facility, SharedUser };