+++ packages/terra-form-input/src/terra-dev-site/doc/example/uncontrolled/DefaultExample.jsx
@@ -1,15 +1,17 @@
 import React from 'react';
-
+import Field from 'terra-form-field';
 // eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
 import Input from 'terra-form-input';
 
 const DefaultExample = () => (
   <div>
-    <h4>Default</h4>
-    <Input
-      name="default input"
-      defaultValue="Default Input – Uncontrolled"
-    />
+    <Field label="Default" htmlFor="uncontrolled-default">
+      <Input
+        name="default input"
+        id="uncontrolled-default"
+        defaultValue="Default Input – Uncontrolled"
+      />
+    </Field>
   </div>
 );