+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -13,6 +13,7 @@ Unreleased
 * Removed use of componentWillReceiveProps
 * Removed use of componentWillMount
 * Updated component to use injectIntl
+* Updated theme variables
 * Updated site examples to be more accessible.
 
 2.38.1 - (November 20, 2018)