+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -3,6 +3,9 @@ ChangeLog
 
 Unreleased
 ----------
+
+2.39.0 - (January 2, 2019)
+------------------
 ### Added
 * Added missing ar translations
 
@@ -10,6 +13,8 @@ Unreleased
 * Removed use of componentWillReceiveProps
 * Removed use of componentWillMount
 * Updated component to use injectIntl
+* Updated theme variables
+* Updated site examples to be more accessible.
 
 2.38.1 - (November 20, 2018)
 ------------------